Make services implement Service, and add index.ts files for imports
Expectation :
-------------------
Change tsconfig.json, so we can do absolute imports :
  import { Service } from '../../../core/service.model';
can be shortened to :
  import { Service } from '@app/core/service.model';
which can also be shortened to :
  import { Service } from '@app/core';
Sweet ! :)

Reality :
-------
When Typescript is compiled, (at least when using CommonJS module),
we don't use import and exports since Node.js don't support them.
So we have to use require(). So what ?
Well, it's fucking impossible to have absolute import without doing some hacky stuff.
Basically, Node look up on current directory, and then use NODE_PATH to lookup where to go next,
which is by default 'node_modules'.
So the solution, is either :

1/ Creating a symlink between the stuff you want to reference and node_modules
Which is kinda okay, as long as no-one is working on your code with Windows,
and also I'm not sure we can use an alias for absolute stuff (like @app/core instead of /core)

2/ Appending NODE_PATH with the root you want
It can be okay also, but the syntax is different between Windows and POSIX systems,
and also we won't be able to create an alias like @app/core

3/ Adding your files directly in node_modules, and adding them to your VCS
No. Just no. I don't want to un-exclude this folder from Intelliji suggestions,
and I shudder at the idea of having to search your file / grep your stuff in there D:

Maybe we could do something like creating a Gulp plugin that read the "paths" stuff,
and then seek the import using the alias, to replace the absolute path by a relative one,
adding as much ../../../../../../ as needed ?
(maybe by enclosing the imports who need to be transformed with special markup in comments ?)
If we do that, do we publish the Gulp plugin on NPM or do we use a private repo (Using Yarn ?)

Still salty about that one, because we where THIS close to have neat imports looking like :
import { Service, ServiceMetadata } from '@app/core';
import { Route, RouteMetadata } from '@app/routing';