Refactor DependencyResolver and fix bad dependency injection
- Rename doYourThing() into resolveDependencies (what a stupid name)

- Put stuff in functions to increase readability of the 2 core methods
(registerModuleServices & resolveDependencies).

- More crucially, fix important bugs in resolveDependency() :
A service were instantiated using dependencies in the order of an array
called resolvedDependenciesId, but this array was constructed
by adding ids only when dependencies where themselves instantiated;
This didn't respected the order given by dependenciesRefs.
Add to that another bug that had resolvedDepenenciesId missing
values when it needed multiple pass to complete,
resulting in undefined passed in a service constructor.
I removed this approach of storing id, and instead stored array of
objects in factory, mutated in-place during passes, but also
much simpler to manipulate.

 (I *really* should try TDD)